requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ML / NLP
transformers==4.40.0
torch==2.2.2
spacy==3.7.4
scikit-learn==1.4.2
xgboost==2.0.3
numpy==1.26.4

# Computer Vision
opencv-python-headless==4.9.0.80
Pillow==10.3.0

# Database
motor==3.4.0
pymongo==4.7.2

# Async task queue (bonus)
celery==5.4.0
redis==5.0.4
flower==2.0.1

# Utils
python-dotenv==1.0.1
pydantic==2.7.1
httpx==0.27.0

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
